+++ lib/events/change.js
@@ -1,7 +1,7 @@
 var UBT = require('../kernel');
-var on = require('../lib/on');
-var parents = require('../lib/parents');
-var collectDataFrom = require('../lib/collect-data-from');
+var on = require('../util/on');
+var parents = require('../util/parents');
+var collectDataFrom = require('../util/collect-data-from');
 
 // 监控值变化事件
 // 逻辑：
