+++ lib/events/route.js
@@ -1,5 +1,5 @@
 var UBT = require('../kernel');
-var on = require('../lib/on');
+var on = require('../util/on');
 
 // 对事件进行管理, 连续操作提交最后一次
 var minimalTimeToRead = 200;
