+++ lib/kernel.js
@@ -1,4 +1,4 @@
-var uuid = require('./lib/uuid');
+var uuid = require('./util/uuid');
 
 // 只有 PV_RECEIVER 才会种植第三方 Cookie
 var NONPV_RECEIVER = 'https://web-ubt.ele.me/par.gif';
