+++ lib/tasks/timing.js
@@ -1,5 +1,5 @@
 var UBT = require('../kernel');
-var on = require('../lib/on');
+var on = require('../util/on');
 
 var timing = function() {
   setTimeout(function() {
