+++ lib/util/collect-data-from.js
